update build_board.py to parallel build
